Update google civic info api to v2
The parsing code is even uglier than it was in the past as we have to
handle multiple ways to figure out what type of official the person
is. Not very happy with it, but even this was fairly painful to get to
as some aspects of the api are not fully supported in the data. Ugh.